[REFACTOR] #93 - BestListAdapter -> BestWrapAdapter로 이름 변경 및 viewType 체크 로직 삭제
